Scene Setup
Added Folders:
-_MyAssets
- _Scenes
- _Scripts
- Animations
- Audio
- Models
- Prefabs

Created a scene called GameScene.  The scene contains a placeholder
player, an invisible plane for the player to jump on, and 3 target
locations for the player to move to (representing the 3 planes of the
game.  Also, 3 prefab placeholder backgrounds were created, they are
placed in the scene where the backgrounds should go.