Create a new obstacle type
A new obstacle type is complete.  This obstacle cannot be jumped over
and is an insta-kill if the player doesn't have the spoon power-up.

The RandomObjectSet class had to be modified slightly because it did not
allow for the spawning of only 1 or 2 obstacles.  This was needed since
the new obstacles needed a way to be avoided.